fix error caused by orphaned released events.
I've turned "Emulate Mouse From Touch" off in the Godot settings and for
some reason, Button controls like to steal/consume the initial touch
pressed event and then ignore the released event. What does this result
in?

Well this causes the InputManager to break because it expects every
pressed to have a matching released. Perhaps this assumption/issue needs
to be sorted out but I've made a quick patch that ignores released
events when there is no matching pressed event.

Because of https://github.com/godotengine/godot/issues/24589
Buttons don't really work with touch so in my project, I've added a
custom script. I noticed @Calinou commented on this issue referring to
Godot Touch Input Manager as a possible solution. Should Godot Touch
Input Manager have some sort of GUI Controls integration that fixes
this? or is this an engine issue? something to think about.